backend/foodgram/requirements.txt
Django==3.2.16
Pillow==9.3.0
djangorestframework==3.14.0
djoser==2.1.0
drf-extra-fields==3.4.1
python-dotenv==0.21.0
gunicorn
psycopg2-binary
django-filter==22.1
